% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChatInterface.R
\name{ChatInterface}
\alias{ChatInterface}
\title{ChatInterface}
\usage{
ChatInterface(fn, ...)
}
\arguments{
\item{fn}{fn The function to wrap the chat interface around. 
Should accept two parameters: a string input message and list of two-element 
lists of the form [[user_message, bot_message], ...] representing the 
chat history, and return a string response.}

\item{...}{Additional arguments, see documentation.}
}
\value{
An gradio Python object.
}
\description{
\url{https://www.gradio.app/docs/gradio/chatinterface}
}
\examples{
\dontrun{
library(gradio)
my_function <- function(message = NULL, history = NULL) {
  return(message)
}

demo <- ChatInterface(fn = my_function, title = "My Chatbot")
demo$launch()
}

}
