% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_gradio.R
\name{install_gradio}
\alias{install_gradio}
\title{Install gradio Python module}
\usage{
install_gradio(method = "auto", conda = "auto")
}
\arguments{
\item{method}{method}

\item{conda}{conda}
}
\value{
(Invisible) `TRUE` if the Conda environment was created successfully.
}
\description{
Install gradio Python module
}
