% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{Interface}
\alias{Interface}
\title{Interface}
\usage{
Interface(fn, inputs, outputs, ...)
}
\arguments{
\item{fn}{fn The function to wrap an interface around. Often a machine learning 
model's prediction function. Each parameter of the function corresponds to 
one input component, and the function should return a single value or a 
tuple of values, with each element in the tuple corresponding to one output 
component.}

\item{inputs}{inputs}

\item{outputs}{outputs}

\item{...}{Additional arguments, see documentation.}
}
\value{
An gradio Python object.
}
\description{
\url{https://www.gradio.app/docs/gradio/interface}
}
\examples{
\dontrun{
library(gradio)
greet <- function(name) {
  return(paste0("Hello, ", name , "!"))
}

demo <- Interface(
  fn=greet,
  inputs= list("text"),
  outputs= list("text"),
)
demo$launch()
}

}
